Compartimentize ChartContainer logic
