+++ packages/x-charts/src/ChartContainer/ChartContainer.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import useForkRef from '@mui/utils/useForkRef';
 import { DrawingProvider, DrawingProviderProps } from '../context/DrawingProvider';
 import {
   SeriesContextProvider,
@@ -8,7 +7,6 @@ import {
 } from '../context/SeriesContextProvider';
 import { InteractionProvider } from '../context/InteractionProvider';
 import { ColorProvider } from '../context/ColorProvider';
-import { useReducedMotion } from '../hooks/useReducedMotion';
 import { ChartsSurface, ChartsSurfaceProps } from '../ChartsSurface';
 import {
   CartesianContextProvider,
@@ -23,7 +21,7 @@ import {
 } from '../context';
 import { ChartsPluginType } from '../models/plugin';
 import { ChartSeriesType } from '../models/seriesType/config';
-import { usePluginsMerge } from './usePluginsMerge';
+import { useChartContainerHooks } from './useChartContainerHooks';
 
 export type ChartContainerProps = Omit<
   ChartsSurfaceProps &
@@ -62,12 +60,14 @@ const ChartContainer = React.forwardRef(function ChartContainer(props: ChartCont
     plugins,
     children,
   } = props;
-  const svgRef = React.useRef<SVGSVGElement>(null);
-  const handleRef = useForkRef(ref, svgRef);
-
-  const { xExtremumGetters, yExtremumGetters, seriesFormatters, colorProcessors } =
-    usePluginsMerge(plugins);
-  useReducedMotion(); // a11y reduce motion (see: https://react-spring.dev/docs/utilities/use-reduced-motion)
+  const {
+    svgRef,
+    handleRef,
+    xExtremumGetters,
+    yExtremumGetters,
+    seriesFormatters,
+    colorProcessors,
+  } = useChartContainerHooks(ref, plugins);
 
   return (
     <DrawingProvider width={width} height={height} margin={margin} svgRef={svgRef}>
